# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Documentation for Furniture App API",
        "description": "Документация API для приложения Furniture App",
        "contact": {
            "email": "support@furnitureapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1/",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/carts/{id}": {
            "get": {
                "tags": [
                    "Корзины"
                ],
                "summary": "Получить корзину по ID",
                "description": "Возвращает корзину по её ID.",
                "operationId": "d0f9c90f921e5cace8d847eb077f6b3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Корзина не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Корзины"
                ],
                "summary": "Обновить корзину",
                "description": "Обновляет существующую корзину по её ID.",
                "operationId": "873541d907e8e905f5ef54604acd7c9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Корзина обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Корзина не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Корзины"
                ],
                "summary": "Удалить корзину",
                "description": "Удаляет существующую корзину по её ID.",
                "operationId": "904192354e6febb57372f82c766440b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Корзина удалена"
                    },
                    "404": {
                        "description": "Корзина не найдена"
                    }
                }
            }
        },
        "/api/carts": {
            "post": {
                "tags": [
                    "Корзины"
                ],
                "summary": "Создать новую корзину",
                "description": "Создаёт новую корзину и возвращает её.",
                "operationId": "06affd160606e43fb60f39628e5bd33a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Cart"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Корзина создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cart"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/cart-items": {
            "get": {
                "tags": [
                    "Элементы корзины"
                ],
                "summary": "Получить все элементы корзины",
                "description": "Возвращает список всех элементов корзины.",
                "operationId": "4233993f8a5847bc1b5329e9d1b898a4",
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Элементы корзины"
                ],
                "summary": "Создать новый элемент корзины",
                "description": "Создаёт новый элемент корзины и возвращает его.",
                "operationId": "0a5df60f7d91780d6e0863dc698a47e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Элемент корзины создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ввод"
                    }
                }
            }
        },
        "/api/cart-items/{id}": {
            "get": {
                "tags": [
                    "Элементы корзины"
                ],
                "summary": "Получить элемент корзины по ID",
                "description": "Возвращает элемент корзины по его ID.",
                "operationId": "cd84f5f7af2553b08417f0c123e80a0f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Операция выполнена успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Элемент корзины не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Элементы корзины"
                ],
                "summary": "Обновить элемент корзины",
                "description": "Обновляет существующий элемент корзины по его ID.",
                "operationId": "78a5883c4ab72212dc40f79a6f5f9232",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CartItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Элемент корзины обновлён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CartItem"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Элемент корзины не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Элементы корзины"
                ],
                "summary": "Удалить элемент корзины",
                "description": "Удаляет существующий элемент корзины по его ID.",
                "operationId": "7ee59aafe2ba3f949091fb7d40d907c7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Элемент корзины удалён"
                    },
                    "404": {
                        "description": "Элемент корзины не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cart": {
                "required": [
                    "id",
                    "user_id",
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "id": {
                        "description": "ID корзины",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID пользователя",
                        "type": "integer"
                    },
                    "product_id": {
                        "description": "ID продукта",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Количество",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CartItem": {
                "title": "CartItem",
                "description": "Cart Item model",
                "required": [
                    "id",
                    "cart_id",
                    "product_id",
                    "quantity"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the cart item",
                        "type": "integer",
                        "format": "int64"
                    },
                    "cart_id": {
                        "description": "ID of the cart",
                        "type": "integer",
                        "format": "int64"
                    },
                    "product_id": {
                        "description": "ID of the product",
                        "type": "integer",
                        "format": "int64"
                    },
                    "quantity": {
                        "description": "Quantity of the product in the cart",
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}